requirements.txt
web3==6.20.1
pandas==2.2.2
matplotlib==3.9.2
python-dotenv==1.0.1
requests==2.32.3
